+++ src/Processor/RedactorProcessorFactory.php
@@ -5,11 +5,13 @@ declare(strict_types=1);
 namespace Sirix\Monolog\Processor;
 
 use Sirix\Monolog\FactoryInterface;
+use Sirix\Monolog\Redaction\Enum\ObjectViewModeEnum;
 use Sirix\Monolog\Redaction\RedactorProcessor;
 
 use function array_key_exists;
 use function is_array;
 use function is_bool;
+use function is_callable;
 use function is_int;
 use function is_string;
 
@@ -46,6 +48,42 @@ class RedactorProcessorFactory implements FactoryInterface
             if (null === $lengthLimit || is_int($lengthLimit)) {
                 $processor->setLengthLimit($lengthLimit);
             }
+        }
+
+        if (isset($options['objectViewMode']) && $options['objectViewMode'] instanceof ObjectViewModeEnum) {
+            $processor->setObjectViewMode($options['objectViewMode']);
+        }
+
+        if (array_key_exists('maxDepth', $options)) {
+            $maxDepth = $options['maxDepth'];
+            if (null === $maxDepth || is_int($maxDepth)) {
+                $processor->setMaxDepth($maxDepth);
+            }
+        }
+
+        if (array_key_exists('maxItemsPerContainer', $options)) {
+            $maxItemsPerContainer = $options['maxItemsPerContainer'];
+            if (null === $maxItemsPerContainer || is_int($maxItemsPerContainer)) {
+                $processor->setMaxItemsPerContainer($maxItemsPerContainer);
+            }
+        }
+
+        if (array_key_exists('maxTotalNodes', $options)) {
+            $maxTotalNodes = $options['maxTotalNodes'];
+            if (null === $maxTotalNodes || is_int($maxTotalNodes)) {
+                $processor->setMaxTotalNodes($maxTotalNodes);
+            }
+        }
+
+        if (array_key_exists('onLimitExceededCallback', $options)) {
+            $callback = $options['onLimitExceededCallback'];
+            if (null === $callback || is_callable($callback)) {
+                $processor->setOnLimitExceededCallback($callback);
+            }
+        }
+
+        if (array_key_exists('overflowPlaceholder', $options)) {
+            $processor->setOverflowPlaceholder($options['overflowPlaceholder']);
         }
 
         return $processor;